Further reorganize artifact structure (#156)
This PR reorganizes the artifact structure. We now have two separate
types of directories to store the libs/weights/..., with one "prebuilt"
directory which holds all the prebuilt libs and weights downloaded from
internet, and other model directories that are generated by local
builds.

CLI and test scripts are updated accordingly for this change.