Allow specifying model name in build
Previously when we introducing the support of "local model path" and
"HuggingFace model path", we removed the support of specifying model
name when build. This turns out preventing us from only specifying the
much shorter model name, as well as doing automatic search for models
existing on the disk.

Therefore, this PR brings back the argparse support for `--model`. Now
we will carefully handle the case where both the model name and one of
the model path / HF path are specified. We now also support model
searching on disk, and support specifying the model only by a short
model name.