Initialize chat module from config JSON string
This PR supports initialization the chat module from JSON-structured
string.
* The CLI is updated with the latest json initialization.
* The iOS/Android app sides are remained using the previous way of
initialization. Now the previous function is renamed to
`init_chat_legacy`, which we will remove after updating the app and web
sides accordingly.
* This PR reverts part of the previous metadata PR as a wrong
assumption was made on the conversation template.
* This PR removes `stream_interval` from chat module as it is unused.
Now `stream_interval` is only kept on the CLI side.