Database in-memory merge
This PR brings the database merge feature, which automatically merges
all database under the specified `--db-path` (which is `./log_db/` by
default).

For example, when there are two database under `./log_db/`:
`./log_db/vicuna` and `./log_db/redpajama`, we won't need to manually
specify one database to use. Instead, `build.py` will now load each
of the database and merge them into one in memory.

We are also backward compatible. Say if we have only one database which
is `./log_db` itself, we are still able to load it in.