LLaMA family loading model on shard level - reducing memory usage (#282)
This PR enables loading LLaMA-family models on shard level. We no longer
use `from_pretrained` to load LLaMA-family models.

This means when loading LLaMA-family models in build.py, we no longer
need to load the full model weights into memory (which is of huge pain
for machines without large enough memory). With this PR, the largest
RAM usage when building LLaMA-family models will be about the size of
the largest shard file.

For example, the largest shard file of Vicuna-7B is roughly 10GB, which
means the largest RAM usage during build is about 10GB.

We implement this functionality based on pass LazyTransformParams and a
topological-sort-based weight computation reordering algorithm.

Corresponding update to GPT-NeoX and RWKV will follow this PR soon.