[iOS][UX] UX Overhaul
This PR overhauls the UX in iOS app. It adds the support for model clear or deletion. And it finalizes the logic of states and transitions in `ModelState` and `ChatState`, with clear and concise code structure. Meanwhile, this PR improves some helper in starting view and alert view for better information displaying.