[Android] Decouple lib and app build
This PR refactors the build of Android application. It decouples the build of `tvm4j` and `mlc-llm` with the build of Android app:

1. prepare the `tvm4j-core.jar` and `libtvm4j_runtime_packed.so`
2. import the lib into the Android app

rather than build the libs and Android app together in the past. Meanwhile, this PR adopts the CMake as the only lib build tool, instead of the past Maven-CMake-NDK combination. With one `prepare_libs.sh`, we are able to build the lib from sources directly with CMake.