Initial `q4f16_1` quantization support with fusion
This PR supports the quantization scheme `q4f16_1` with desired fusion
pattern. Compared with `q3f16_0` or `q4f16_0`,

* this scheme does not transpose the linear layer weights in
quantization,
* under this scheme we fuse all transpose with matmul so that the fused
matmul is of layout "NT".

The purpose of this PR is to provide initial support of the quantization
scheme and ensure the desired fusion pattern under this scheme.
The TIR functions under this scheme is not fully optimized. We will
introduce the optimization with dlight (the default schedule package)
soon.

The performance of other quantization schemes are not affected.
We will do clean up for some of the optimization passes in the future
(for example, remove the "`gemv_only`" condition in decode-matmul-ewise
fusion) when dlight and ParamManager get more mature.