Clean up `op_pattern` attr in all places
Recently, the introduction of ParamManager in #477 changes the graph
fusion pattern in certain level, which may cause TIR mismatch in
the dispatch and/or database application of the build stage: on one
side there exists the attr `op_pattern`, and on the other side there
does not exist the attr now.

Though some mismatches in dispatch/database application can be easily
noticed due to rapid performance dropdown, some other mismatches might
be hard to catch. These mismatches may contribute to performance
degradation alltogether.

Therefore, this PR introduces a pass that cleans up the `op_pattern`
attrs in the IRModule. This PR also removes all the `op_pattern` attrs
in existing databases and dispatches.

With this PR, we no longer need to worry about TIR mismatch due to
atttributes.