[Fix] FuseDecodeTranspose pass with PrimFunc deep copy
Prior to this PR, the FuseDecodeTranspose pass creates the fused
PrimFunc which directly reuses the blocks in the functions being fused.
This lead to tir.Schedule error when a Schedule is constructed on an
IRModule which contains both the post-fusion and pre-fusion functions.

This PR calls `tir.stmt_functor.renew_defs` on the fused PrimFunc for
the purpose of simple deep copy.

Other fusion passes are confirmed with no similar issue.