[Refactor] Param loading integration into ParamManager
This PR refactors the lazy parameter loading logic, integrating it
into the ParamManager so that the concept of ParamManager is more
self-contained.

On top of this, this PR supports the lazy parameter loading with an
additional functionality. Prior to this PR, the lazy parameter
loading only supports the case of "one Relax parameter comes from one
torch parameter". In this setting, we can support "the torch nn.Module
has combined QKV while our Relax nn.Module separates them". However,
we did not support the reverse - we did not support "the torch nn.Module
separates Q, K, V, and the Relax nn.Module combines them together".

This PR now supports this. It opens the floor for the QKV combination
for models like LLaMA (whose PyTorch nn.Module implementation does not
combine Q, K, V together).

Verified the build on all models.