Lift weight conversion to an early stage
This main part of this PR separates the weight conversion func
construction from the model IRModule transformation, so that we can now
choose to only convert the model weights or only to build the model
library, by specifying corresponding build arguments.

With this PR, the `mod_transform_before_build` function no longer
has anything to do with the weight conversion computation.

Besides, this PR also
* removes some dump functions to utils,
* making the dumped `.pkl` IRModule independent of the architecture.