Supports more general quantization func registration (#622)
This PR supports the quantization function registration of a
QuantizationSpec with a relax function that takes in a BlockBuilder and
a list of relax Vars, computes the (de)quantization and returns the
results.

Prior to this PR, only TE functions are allowed, which is a bit
restrictive. With this function, we are able to register more general
quantization computation (such as calling external functions), which
enables more possibility.