Deprecate MetaSchedule database
This PR formally disables the application of MetaSchedule database.

Confirmed that the deprecation does not bring performance regression
to both Llama2-7b and RedPajama-3b on both RTX-4090 with CUDA and
M1 Max MacStudio. (It brings better performance to RedPajama-3b as some
tuned workloads in the database are less optimal.)