[ChatModule] Raw text generation and benchmark (#697)
This PR introduces the `RawGenerate` function to ChatModule for
benchmarking purpose, which generates text by taking the input prompt
and target generation length. The raw text generation does not use
system prompt and ignores the stop tokens of the conversation template.

Based on the raw generation, this PR also brings the
`benchmark_generate` method to the Python ChatModule. The function
first warms up and then invoke the raw generation. After
`benchmark_generate`, we can use `stats()` to get the generation
performance of the model.

The `benchmark_generate` methods replaces the previous public
`evaluate` method in the Python ChatModule.