[Llama] Online rotary embedding
This PR supports Llama2 with the online rotary embedding computation
and removes the dependency on the cached cos/sin values, so that
no matter how much the max-seq-len of the newly built libraries
is, the prebuilt weights can be used without issues (hopefully).

Tested `q4f16_1` on RTX 4090 and observed no performance regression.