[BugFix] Revert Breaking Change to `ChatModule.Reload` API
In multi-GPU setting, a hard constraint is that currently only shared
library is supported, and a string path is required to be provided to
Disco's session as:

```python
sess.load_vm_module("/path/to/shared/lib.so")
```

And therefore, previously in PR #867, we made a breaking change to
the signature of `ChatModule.reload` API as below:

```diff
- void Reload(tvm::runtime::Module lib, ...);
+ void Reload(String lib_path, ...);
```

However, it brings an unexpected breaking change, which on iOS and
Android apps, we expect the library being bundled together as static
system libraries rather than shared libraries, and in these cases,
the string path is not available, making them crash after the PR merge.

This PR re-enables backward compatibility for those apps and also
preserves the ability to run multi-GPU in TVM Unity's Disco runtime by
allowing a union type:

```diff
- void Reload(String lib_path, ...);
+ void Reload(Union[tvm::runtime::Module, String] lib_path, ...);
```

Tracking issue: https://github.com/mlc-ai/mlc-llm/issues/936.