[Llama] Rename dynamic vocab size from "v" to "vocab_size"
Single-letter variable names may be difficult to recognize when
encountered in other contexts.  This commit changes the name of the
`tir.Var` used for the vocab size in the llama model definition, so
that it can be more easily recognized when occurring in error messages
and when used in `relax.transform.BindSymbolicVars`.