[Core][Llama] Argument `max_vocab_size` and `max_batch_size`
This PR introduces the `max_vocab_size` and `max_batch_size`
as two new compile arguments. The purpose is for better
memory planning.

Besides, this PR updates llama to make use of the two arguments.
Other models are not changed yet.

The default value for `max_vocab_size` is set to 40000, which
I think is larger than the values of most models. The default
value for `max_batch_size` is currently set as 256. It is possible
that we update this value in the future to have a good default
number.