[Serving] Serving Engine (#1075)
This PR introduces the engine for request serving.

The PR includes the C++ side (which is the main implementation
of the engine) and the Python side (which is a wrapper class of
Engine in C++). This is similar to our implementation of
ChatModule.

One test file is added to run text generation for a batch of
sequences by directly calling into the Engine methods.