[Serving] Update sampling logits to unify memory copy (#1187)
This PR unifies the memory copy (GPU to CPU) of logits/prob
arrays. Prior to this PR, the logits/probs array of each token
is copied separately. This incurs some overhead without doubt.
This PR unifies the memory copy, so that we only need to copy
once for the entire running batch during decoding at a time.

In terms of implementation, this PR simplies the sampler logic
by completely removing the memory copy logic. On the other hand,
it brings the copy logic to engine. It does not bring much
complexity to the engine.

This PR also lifts the random number generation from sampler to
engine, as a preparation step for potential multi-threading
token sampling.