[Serving] Support multi-threading CPU sampling
This PR supports the multi-threading token sampling on CPU.

In serving scenarios, the token sampling process becomes one of the
bottlenecks, as the model computation has higher throughput than
single-sequence settings. Therefore, we enhances the CPU sampling
with multi-threading.

Particularly,
* this PR changes the design scope of Sampler. Prior to this PR,
the sampling function exposed by Sampler focuses on sampling a
single token. After this PR, a function processes a batch of
tokens. This makes the multi-threading more manageable.
* the multi-threading at this moment is backed by OpenMP, according
to our micro-benchmark.

Note: to effectively enable OpenMP, now need to compile with gcc/g++.