[Serving][Refactor] Major codebase refactor
This PR is a major refactor of the serving framework.
It contains the following aspects:

* Changing Model and Sampler from `runtime.Module` to Object
in TVM. Exposing the public interface through public member
function (rather than through `GetFunction` which returns PackedFunc).
Separating the definition and implementation of Model/Sampler
classes. Now a base Model/Sampler class definition is in their
respective header files, and the implementations are in .cc files.
* Removing the TokenizerModule class, and directly using the
Tokenizer class in `tokenizer_cpp`, to reduce indirection.
* Introducing unique string `id` to Request. This id is passed
in from frontend as a Request constructor parameter, and is the
unique identifier of a request.
* Reducing the uses of `ShapeTuple` after the Model/Sampler/Tokenizer
interface changes.
* Moving some previous member functions in Engine (such as "getting
data length", "getting data embedding") to the corresponding
data structure side.
* Introducing struct `EngineStats` to contain all the runtime
statistics of the engine.
* Classifying and reordering the member functions in Engine.