[Serving][Refactor] EngineState and EngineAction
Following #1277, this PR continues to refactor the engine.

Particularly, this PR introduces the `EngineState` object
to denote the running state of an Engine, and the `EngineAction`
abstraction to denote the action (such as prefill, decode,
speculate, etc.) that the engine can take at each time step.

The EngineState consists of
* the queues of running requests, waiting requests and the requests
to be aborted,
* the request states of all requests,
* the engine runtime statistics.

The EngineAction contains the core interface
```c++
bool Step(EngineState estate)
```
which takes the current engine state as input and then update
the engine state after taking the action. With the introduction
of EngineAction, we now further move a bunch of detailed implementation
functions under Engine to their respective actions.