[Serving] Refactor Engine Logics
This PR refactors the engine logics with the following changes:

1) Adopt explicit [vtable convention](https://github.com/mlc-ai/mlc-llm/blob/a40d0cc15d7fe2e64ea8cc68b543ffa557301889/cpp/serve/engine.cc#L202-L209)
of TVM's runtime Module in C++;
2) Migrate to vtable convention in Python by consolidating vtable
functions into a single `self._ffi` member as the Python-side vtable,
created by a single [TVM call](https://github.com/mlc-ai/mlc-llm/blob/a40d0cc15d7fe2e64ea8cc68b543ffa557301889/python/mlc_chat/serve/engine.py#L120-L123).
3) Consolidate and unify automatic device detection logic into
`mlc_chat.support.auto_device` with detailed logging logic and fallback
mechanisms.
4) Explicitly require `max_single_sequence_length` and `tokenizer_path`
on C++ side, and to keep backward compatibility, those two arguments are
now automatically detected in Python;
5) Dissolve the `_reload` method into `__init__` (Python) and `init`
(C++).
6) Reduce the complexity of Engine creation logic, which now takes 3
major steps (was 5).

There are still a few items to be refactored, particularly around model
and function table, for example, model creation logic now uses a
dynamically typed `model_lib`, and it along could not decide if the
model belongs to Disco (multi-GPU distributed runtime) or a normal
runtime, and the function table that (un)-naturally comes with this
limitation.

Another thing to look at is that the creation of temporary buffers, as
well as the data transfer in-between are not sufficiently managed at
first glance. There seems to be a few `NDArray::Empty` and
`NDArray::CopyFromTo` calls in the sampler as well as the model.
`FunctionTable::Empty` exists as well but is not yet used - but it will
become suspicious as soon as being used.

Threading model of this engine is still a bit basic, but I think it's
pretty much future proof to potential extensions at the moment.

The last important note - this PR depends on a couple of changes on TVM
upstream:
- https://github.com/apache/tvm/pull/16147
- https://github.com/apache/tvm/pull/16148

Before synching TVM Unity and this repo, to compile successfully, one
has to put the line below in `build/config.cmake`:

```cmake
set(TVM_HOME /path/to/tvm/unity)
```