[Python] Detect Driver/Device in a Separate Process (#1311)
This PR separates the device detection into separate
subprocesses.

The change is because the device detection will setup the
driver, which consumes some GPU VRAM (for example,
`tvm.device("opencl", 0).exist` consumes 390MB of VRAM on
RTX 4090). Consider the case we detect if CUDA, Vulkan and
OpenCL are available. When they are all available, each
detection holds some VRAM, larger than 430MB altogether.

If the device detection is in the same process as the main
process, the VRAM consumed by device detection will never
be released. This means that in the example above, we
detect CUDA, Vulkan and OpenCL, while in the end we
prioritize the CUDA device. Consequently, the memory held by
Vulkan and OpenCL detection will never be released.

Motivated by this issue, we separate the detection into
subprocess, so that the held VRAM can be successfully
released after detection.