[Serving] TextStreamer and StopStringHandler
This PR introduces the TextStreamer and StopStringHandler
as helper classes of tokenizers, and updates Engine, AsyncEngine
and AsyncThreadedEngine accordingly to make use of the helper
classes to handle stop strings and reduce detokenization overhead.

Specifically,
* `TextStreamer` accepts a list of delta token ids one at a time,
and return a delta text string which is guaranteed to be UTF-8 valid.
With `TextStreamer`, we can properly stream back generated texts
to frontend users without worrying about returning strings that
are not UTF-8 valid.
* `StopStringHandler` is aware of the stop strings of a given
request, accepts a piece of input delta text string each time,
check if any stop string is hit, and return a piece of output
delta text that is guaranteed not to be part of the stop string.
With `StopStringHandler`, we can properly handle stop strings.

For tests,
* this PR adds the pytest-compatible tests for `TextStreamer`
and `StopStringHandler` in file
`tests/python/support/test_text_streamer_stop_string_handler.py`.
* this PR adds the server tests of stop strings, and the test
for external cancellation (implemented via request timeout).

Note that we offload all detokenization to Python side, and
the stop string handling is not part of Engine implementation
in C++.