[Serving] Conversation template and OpenAI v1/chat/completions
This PR supports the `v1/chat/completions` entrypoint of OpenAI
API:
* protocol: `python/mlc_chat/protocol/openai_api_protocol.py`,
* entrypoint: `python/mlc_chat/serve/entrypoints/openai_entrypoints.py`

To support the chat functionality, this PR introduces the conversation
template definition and register on Python side in
`python/mlc_chat/serve/conversation.py`.
Compared with the current implementation of conversation template
in C++, this version supports overriding the content of the system
prompt without changing formatting texts (like the `[INST]` string
in Llama2).

This PR adds a series of tests:
* temperature test for `v1/completions`,
* three different history message tests for `v1/chat/completions`,
* `max_tokens` test for `v1/chat/completions`,
* "wrong system prompt" position test for `v1/chat/completions`,
* the `openai` package test for both `v1/completions` and
`v1/chat/completions`.