[Serving] Introducing EventTraceRecorder
This PR introduces the EventTraceRecorder in MLC LLM serving, which can
record event times for each request on invocation, and can dump
all the recorded events to Chrome Trace Event format, so that
we can visualize the events and the timeline of all requests.

This PR brings the optional flag `enable_tracing` to Engine,
AsyncEngine/AsyncThreadedEngine, server entry and PopenServer.
This flag defaults to `False`, in which case we do not enable
tracing.

This PR introduces a new server entrypoint `/debug/dump_event_trace`,
under file `python/mlc_chat/serve/entrypoints/debug_entrypoints.py`.
This entrypoint accepts a model name, and returns the dumped events
in Chrome Trace Event format (in JSON string).

This PR contains a basic EventTraceRecorder test and a basic test for
entrypoint `/debug/dump_event_trace`.