[Serving] Fix AsyncThreadedEngine for Metal
This PR fixes a threading issue when using AsyncThreadedEngine
on Metal. This issue does not exist for CUDA.

Prior to this PR, the engine is constructed on one threaded
and is driven by the other thread. After this PR, both the
construction and the engine drive is on the same thread, which
does not raise issue in the current Metal runtime.