[Serving][Grammar] BNF AST and Parser for EBNF grammar (#1534)
This PR brings the first part of the support of grammar guided
generation. The whole pipeline is:

1. User provide a grammar in BNF or EBNF format
2. Parse the grammar into a AST
3. Simplify the grammar. Notably, eliminate the left recursion.
4. Match the syntax online and obtain the possible values of the next
token as a mask
5. Use the mask to generate the next token

This PR achieves 1. and 2. It contains the design of the AST of BNF
grammar. For the AST description, see `serve/grammar/grammar.h`. 

Also, this PR implements a (almost) LL(1) parser to parse and simplify
the given grammar source file into a BNF AST.

A printer of the AST is provided for debug purpose.