[Serving] Automatically estimate KV cache capacity
This PR supports the functionality that automatically estimates
the maximum KV cache capacity. This leverages the following
sources:
* total GPU global memory (through TVM device API),
* model parameter memory (computed during weight conversion),
* runtime temporary memory (comptued by memory planning).

With this PR, when `max_total_sequence_length` is set to None,
we will automatically estimate a KV cache capacity that can
make most use of the GPU.

Alternatively, users can choose to specify total GPU global
memory size via the environment variable `"MLC_GPU_SIZE_BYTES"`.
If this env variable is detected, we will use this number as
the GPU VRAM size.

Effects of this PR:
```
[2024-01-14 15:35:39] INFO engine.py:186: Estimated KVCacheConfig "max_total_sequence_length": 18267.
[2024-01-14 15:35:39] INFO engine.py:191: Estimated total single GPU memory usage: 23374.68 MB (Parameters: 12852.51 MB. KVCache: 9590.18 MB. Temporary buffer: 932.00 MB)
```