[CI] Test Model Compilation
This PR adds basic testing infra for the command:

```bash
mlc_chat chat ...
```

This is imminently necessary because it got broken quite frequently upon
almost every new feature, and it's quite non-trivial to test all the
quadruples `(model, quant, device, tensor-parallel-shards)`.

This testing infra addresses this issue by installing the wheel produced
by the previous build pipeline, and run `mlc_chat chat` commands in a
subprocess to make sure at least the compilation process will not break.
Correctness is not validated in this PR, as they are more likely to be
guaranteed by unit-tests.

This PR is largely inspired by @Hzfengsy's series of work, particularly
the set up of Jenkins pipeline.

Co-authored-by: Siyuan Feng <Hzfengsy@sjtu.edu.cn>